backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
requests>=2.32.0
beautifulsoup4>=4.13.0
llama-index-core>=0.13.0
python-dotenv>=1.0.1
pytest>=8.3.0
